fix integration fixture setup
